requirements.txt
asgiref==3.2.3
autopep8==1.5
awesome-slugify==1.6.5
Django==3.0.4
django-appconf==1.0.4
django-cors-headers==3.2.1
django-imagekit==4.0.2
djangorestframework==3.11.0
pilkit==2.0
Pillow==7.1.1
psycopg2==2.8.5
pycodestyle==2.5.0
pytz==2019.3
regex==2020.2.20
six==1.14.0
sqlparse==0.3.1
Unidecode==0.4.21
